requirements.txt
h5py
tensorflow-gpu==1.4.0
Keras==2.0.8